+++ Assets/Scripts/Simulator/VectorQuantity.cs
@@ -4,6 +4,6 @@ namespace AircraftSimulator {
     public abstract class VectorQuantity : Quantity {
         protected VectorQuantity(Model model) : base(model) { }
 
-        public abstract Vector3 Value(Vector3 position, double time);
+        public abstract Vector3 Value(Vector3 position, float time);
     }
 }
