+++ Assets/Scripts/Simulator/Wind.cs
@@ -11,7 +11,7 @@ namespace AircraftSimulator {
             _model = model;
         }
 
-        public override Vector3 Value(Vector3 position, double time) {
+        public override Vector3 Value(Vector3 position, float time) {
             if (_model is ConstantWindModel constantWindModel) return constantWindModel.Value;
             if (_model is TurbulentWindModel turbulentWindModel) return turbulentWindModel.Value(position, time);
             throw new Exception("Invalid wind model");
